requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyPDF2
language-tool-python
openai
pyspellchecker
